move term to kubectl/util
