+++ pkg/kubectl/cmd/BUILD
@@ -94,6 +94,7 @@ go_library(
         "//pkg/kubectl/plugins:go_default_library",
         "//pkg/kubectl/resource:go_default_library",
         "//pkg/kubectl/util:go_default_library",
+        "//pkg/kubectl/util/term:go_default_library",
         "//pkg/kubelet/types:go_default_library",
         "//pkg/printers:go_default_library",
         "//pkg/printers/internalversion:go_default_library",
@@ -101,7 +102,6 @@ go_library(
         "//pkg/util/i18n:go_default_library",
         "//pkg/util/interrupt:go_default_library",
         "//pkg/util/taints:go_default_library",
-        "//pkg/util/term:go_default_library",
         "//pkg/version:go_default_library",
         "//vendor/github.com/daviddengcn/go-colortext:go_default_library",
         "//vendor/github.com/docker/distribution/reference:go_default_library",
@@ -214,11 +214,11 @@ go_test(
         "//pkg/kubectl/cmd/util/openapi:go_default_library",
         "//pkg/kubectl/plugins:go_default_library",
         "//pkg/kubectl/resource:go_default_library",
+        "//pkg/kubectl/util/term:go_default_library",
         "//pkg/printers:go_default_library",
         "//pkg/printers/internalversion:go_default_library",
         "//pkg/util/i18n:go_default_library",
         "//pkg/util/strings:go_default_library",
-        "//pkg/util/term:go_default_library",
         "//vendor/github.com/go-openapi/spec:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
