+++ pkg/kubectl/cmd/templates/BUILD
@@ -19,7 +19,7 @@ go_library(
         "//build/visible_to:pkg_kubectl_cmd_templates_CONSUMERS",
     ],
     deps = [
-        "//pkg/util/term:go_default_library",
+        "//pkg/kubectl/util/term:go_default_library",
         "//vendor/github.com/MakeNowJust/heredoc:go_default_library",
         "//vendor/github.com/russross/blackfriday:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
