+++ pkg/kubectl/cmd/templates/templater.go
@@ -23,7 +23,7 @@ import (
 	"text/template"
 	"unicode"
 
-	"k8s.io/kubernetes/pkg/util/term"
+	"k8s.io/kubernetes/pkg/kubectl/util/term"
 
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
