+++ pkg/kubectl/cmd/util/editor/BUILD
@@ -22,8 +22,8 @@ go_library(
         "//pkg/kubectl/cmd/util:go_default_library",
         "//pkg/kubectl/resource:go_default_library",
         "//pkg/kubectl/util/crlf:go_default_library",
+        "//pkg/kubectl/util/term:go_default_library",
         "//pkg/printers:go_default_library",
-        "//pkg/util/term:go_default_library",
         "//vendor/github.com/evanphx/json-patch:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
